crosvm: linux: allow video device to be created even if /usr/lib64 does not exist
`create_video_device` currently fails if `/usr/lib64` does not exist on
the host, because it cannot be bound to the jail in that case. This
should not be a fatal error as this directory is only needed for AMD
devices and may not exist on all hosts.

In case this directory does not exist on the host, the video device
process will just exit with an obscure error code as minijail's
`process_mounts_or_die` cannot mount it. Logging is not visible to the
user for some reason, so this makes it a pretty difficult issue to
debug.

Fix this by extracting the code that mounts existing directories in
`gpu_jail` and using it from `create_gpu_device` as well.

BUG=b:210599385
TEST=crosvm can start with `--video-decoder` on a host where
`/usr/lib64` does not exist.

Change-Id: Iacdfe6d3925689ba645c07d6cdbcbab5e1729b5c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crosvm/+/3341133
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Dennis Kempin <denniskempin@google.com>
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>