Use cfg_if crate to disambiguate the chosen VideoBackendType
The logic for the None branch quickly becomes verbose as more video backends
are added. In preparation for the new libva backend, use the cfg_if crate
to disambiguate the chosen VideoBackendType when the user does not specify one.

BUG=b:214478588
TEST="cargo-expand --bin crosvm --features=video-decoder,vaapi,ffmpeg -- parse_video_options
shows a single None branch"

Change-Id: I0b7489924c30b2834cecc636446ac96034f6adc5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crosvm/+/3422776
Tested-by: kokoro <noreply+kokoro@google.com>
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>
Commit-Queue: Alexandre Courbot <acourbot@chromium.org>